pyproject.toml
[project]
name = "lang-chain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "black>=25.12.0",
    "fastapi[standard]>=0.124.0",
    "isort>=7.0.0",
    "langchain>=1.1.2",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.0",
    "langchain-tavily>=0.2.13",
    "langgraph>=1.0.4",
    "langgraph-checkpoint-postgres>=3.0.2",
    "psycopg>=3.3.2",
    "psycopg-binary>=3.3.2",
    "psycopg2-binary",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]

[tool.setuptools.packages.find]
where = ["src"]
